saves user
when the user enters in their name, it now saves as a new user along with the user Id. Next i need to create the actual channel page and functionality and its turn-in ready. Debating if i want to add a create channel method so the user can create their own channels. But i'll decide that after.