basically done but there are a few bugs
The names are being sent as strings in the conversation box, so i need to fix that. Also i want to clear the input box once a message is sent and possibly add an event listener for the enter key. if i'm feeling crazy i'll create a method to manually create channels.